backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
openai==1.72.0
pydantic==2.11.3
pydantic-settings==2.8.1
websockets==15.0.1